+++ server/src/main/java/io/spine/server/BoundedContextBuilder.java
@@ -45,7 +45,6 @@ import io.spine.server.entity.Repository;
 import io.spine.server.event.EventBus;
 import io.spine.server.event.EventDispatcher;
 import io.spine.server.event.EventEnricher;
-import io.spine.server.integration.IntegrationBroker;
 import io.spine.server.stand.Stand;
 import io.spine.server.tenant.TenantIndex;
 import io.spine.server.type.CommandEnvelope;
@@ -90,7 +89,7 @@ public final class BoundedContextBuilder implements Logging {
 
     /**
      * Event dispatchers to be registered with the context {@link EventBus} and/or
-     * {@link IntegrationBroker} after the Bounded Context creation.
+     * {@link io.spine.server.integration.IntegrationBroker} after the Bounded Context creation.
      */
     private final Collection<EventDispatcher> eventDispatchers = new ArrayList<>();
