+++ server/src/main/java/io/spine/server/integration/DomesticEventPublisher.java
@@ -39,7 +39,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * a remote channel.
  *
  * <p>The events to subscribe are those that are required by external application components
- * at this moment; their set is determined by the {@linkplain RequestForExternalMessages
+ * at this moment; their set is determined by the {@linkplain ExternalEventsWanted
  * configuration messages}, received by this instance of {@code IntegrationBroker}.
  */
 final class DomesticEventPublisher implements EventDispatcher, Logging {
