+++ server/src/test/java/io/spine/server/integration/DefaultIntegrationBrokerTest.java
@@ -51,8 +51,8 @@ import static io.spine.server.integration.given.broker.IntegrationBrokerTestEnv.
 import static io.spine.server.integration.given.broker.IntegrationBrokerTestEnv.subscribedStatisticsBc;
 import static io.spine.server.integration.given.broker.IntegrationBrokerTestEnv.subscribedWarehouseBc;
 
-@DisplayName("`IntegrationBroker` should")
-class IntegrationBrokerTest {
+@DisplayName("`DefaultIntegrationBroker` should")
+class DefaultIntegrationBrokerTest {
 
     @BeforeEach
     void setUp() {
